+++ util.c
@@ -1,7 +1,7 @@
 #include <stdarg.h>
 #include "mle.h"
 
-int util_file_exists(char* path, size_t path_len) {
+int util_file_exists(char* path, int path_len) {
     // TODO util_file_exists
     return 0;
 }
